Fix: AptPresent check was not working on specific package due to broken regex
Solution: Removed the \s+ after the version number in regex, in case no version was provided it forced the package name to have at least two whitespaces between name and version which is not the case in case of long package name.

A unit test check if the specific problem still appears.